+++ macros/get_invoice_columns.sql
@@ -41,6 +41,7 @@
     {"name": "status_transitions_marked_uncollectible_at", "datatype": dbt_utils.type_timestamp()},
     {"name": "status_transitions_paid_at", "datatype": dbt_utils.type_timestamp()},
     {"name": "status_transitions_voided_at", "datatype": dbt_utils.type_timestamp()},
+    {"name": "subscription_id", "datatype": dbt_utils.type_string()},
     {"name": "subscription_proration_date", "datatype": dbt_utils.type_int()},
     {"name": "subtotal", "datatype": dbt_utils.type_int()},
     {"name": "tax", "datatype": dbt_utils.type_int()},