+++ CHANGELOG.md
@@ -1,7 +1,12 @@
 # dbt_stripe_source v0.7.1
+## 🎉 Documentation and Feature Updates
 - Add timezone notation to timestamp column definition ([#42](https://github.com/fivetran/dbt_stripe_source/pull/42))
+- Addition of subscription ID in the `stg_stripe__invoice` model ([#41](https://github.com/fivetran/dbt_stripe_source/pull/41))
 
+## Contributors
 - [mvtemple](https://github.com/mvtemple) ([#43](https://github.com/fivetran/dbt_stripe_source/issues/43))
+- [ashishk-de](https://github.com/ashishk-de) ([#44](https://github.com/fivetran/dbt_stripe_source/issues/44))
+
 # dbt_stripe_source v0.7.0
 ## 🎉 Documentation and Feature Updates
 - Updated README documentation updates for easier navigation and setup of the dbt package